pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lakeview-dashboard"
version = "0.1"
description = "A dashboard management tool for Databricks Lakeview"
authors = [{name = "Rohit Bhagwat", email = "rohit.bhagwat@databricks.com"}]
dependencies = [
    "databricks-sdk",
    "python-dotenv",
    "pydantic",
]

[project.optional-dependencies]
dev = ["pytest"]